# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Gorax - a powerful workflow automation platform with webhooks, scheduling, and real-time execution monitoring.\n\n## Authentication\nAll API endpoints (except /health, /ready, and webhook triggers) require authentication.\nIn development mode, use the X-User-ID header. In production, use Ory Kratos session cookies.\n\n## Multi-tenancy\nAll authenticated endpoints require a valid X-Tenant-ID header to identify the tenant context.\n\n## Rate Limiting\nAPI requests are rate-limited based on tenant quotas. Check response headers for limit information.\n\n## Error Handling\nErrors follow a consistent format with appropriate HTTP status codes and error messages.",
        "title": "Gorax Workflow Automation API",
        "contact": {
            "name": "Gorax Support",
            "url": "https://github.com/gorax/gorax",
            "email": "support@gorax.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/analytics/errors": {
            "get": {
                "security": [
                    {
                        "TenantID": []
                    },
                    {
                        "UserID": []
                    }
                ],
                "description": "Returns categorized error data including error types, frequencies, and affected workflows",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get error breakdown and analysis",
                "parameters": [
                    {
                        "type": "string",
                        "example": "2024-01-01T00:00:00Z",
                        "description": "Start date (RFC3339 format)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2024-01-31T23:59:59Z",
                        "description": "End date (RFC3339 format)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Error breakdown data",
                        "schema": {
                            "$ref": "#/definitions/github_com_gorax_gorax_internal_analytics.ErrorBreakdown"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/analytics/overview": {
            "get": {
                "security": [
                    {
                        "TenantID": []
                    },
                    {
                        "UserID": []
                    }
                ],
                "description": "Returns aggregated analytics data for the tenant including total executions, success rates, and trends",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get tenant analytics overview",
                "parameters": [
                    {
                        "type": "string",
                        "example": "2024-01-01T00:00:00Z",
                        "description": "Start date (RFC3339 format)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2024-01-31T23:59:59Z",
                        "description": "End date (RFC3339 format)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tenant analytics overview",
                        "schema": {
                            "$ref": "#/definitions/github_com_gorax_gorax_internal_analytics.TenantOverview"
                        }
                    },
                    "400": {
                        "description": "Invalid time range",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/analytics/top-workflows": {
            "get": {
                "security": [
                    {
                        "TenantID": []
                    },
                    {
                        "UserID": []
                    }
                ],
                "description": "Returns the most frequently executed workflows ordered by execution count",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get top workflows by execution count",
                "parameters": [
                    {
                        "type": "string",
                        "example": "2024-01-01T00:00:00Z",
                        "description": "Start date (RFC3339 format)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2024-01-31T23:59:59Z",
                        "description": "End date (RFC3339 format)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Maximum number of workflows to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Top workflows",
                        "schema": {
                            "$ref": "#/definitions/github_com_gorax_gorax_internal_analytics.TopWorkflows"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/analytics/trends": {
            "get": {
                "security": [
                    {
                        "TenantID": []
                    },
                    {
                        "UserID": []
                    }
                ],
                "description": "Returns time-series data showing execution trends with configurable granularity (hour, day, week, month)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get execution trends",
                "parameters": [
                    {
                        "type": "string",
                        "example": "2024-01-01T00:00:00Z",
                        "description": "Start date (RFC3339 format)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2024-01-31T23:59:59Z",
                        "description": "End date (RFC3339 format)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "hour",
                            "day",
                            "week",
                            "month"
                        ],
                        "type": "string",
                        "default": "day",
                        "description": "Time granularity (hour, day, week, month)",
                        "name": "granularity",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Execution trends data",
                        "schema": {
                            "$ref": "#/definitions/github_com_gorax_gorax_internal_analytics.ExecutionTrends"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/analytics/workflows/{workflowID}": {
            "get": {
                "security": [
                    {
                        "TenantID": []
                    },
                    {
                        "UserID": []
                    }
                ],
                "description": "Returns detailed analytics for a specific workflow including execution counts, success rate, and average duration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get workflow-specific analytics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workflow ID",
                        "name": "workflowID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2024-01-01T00:00:00Z",
                        "description": "Start date (RFC3339 format)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2024-01-31T23:59:59Z",
                        "description": "End date (RFC3339 format)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Workflow analytics",
                        "schema": {
                            "$ref": "#/definitions/github_com_gorax_gorax_internal_analytics.WorkflowStats"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Workflow not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/analytics/workflows/{workflowID}/nodes": {
            "get": {
                "security": [
                    {
                        "TenantID": []
                    },
                    {
                        "UserID": []
                    }
                ],
                "description": "Returns performance statistics for individual nodes within a workflow including execution time and success rate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get node performance metrics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workflow ID",
                        "name": "workflowID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Node performance metrics",
                        "schema": {
                            "$ref": "#/definitions/github_com_gorax_gorax_internal_analytics.NodePerformance"
                        }
                    },
                    "400": {
                        "description": "Invalid workflow ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Workflow not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_gorax_gorax_internal_analytics.ErrorBreakdown": {
            "type": "object",
            "properties": {
                "errorsByType": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_gorax_gorax_internal_analytics.ErrorInfo"
                    }
                },
                "totalErrors": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_gorax_gorax_internal_analytics.ExecutionTrends": {
            "type": "object",
            "properties": {
                "dataPoints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_gorax_gorax_internal_analytics.TimeSeriesPoint"
                    }
                },
                "endDate": {
                    "type": "string"
                },
                "granularity": {
                    "$ref": "#/definitions/github_com_gorax_gorax_internal_analytics.Granularity"
                },
                "startDate": {
                    "type": "string"
                }
            }
        },
        "github_com_gorax_gorax_internal_analytics.Granularity": {
            "type": "string",
            "enum": [
                "hour",
                "day",
                "week",
                "month"
            ],
            "x-enum-varnames": [
                "GranularityHour",
                "GranularityDay",
                "GranularityWeek",
                "GranularityMonth"
            ]
        },
        "github_com_gorax_gorax_internal_analytics.NodePerformance": {
            "type": "object",
            "properties": {
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_gorax_gorax_internal_analytics.NodeStats"
                    }
                },
                "workflowId": {
                    "type": "string"
                },
                "workflowName": {
                    "type": "string"
                }
            }
        },
...
        "github_com_gorax_gorax_internal_analytics.TenantOverview": {
            "type": "object",
            "properties": {
                "activeWorkflows": {
                    "type": "integer"
                },
                "avgDurationMs": {
                    "type": "integer"
                },
                "cancelledExecutions": {
                    "type": "integer"
                },
                "failedExecutions": {
                    "type": "integer"
                },
                "pendingExecutions": {
                    "type": "integer"
                },
                "runningExecutions": {
                    "type": "integer"
                },
                "successRate": {
                    "type": "number"
                },
                "successfulExecutions": {
                    "type": "integer"
                },
                "totalExecutions": {
                    "type": "integer"
                },
                "totalWorkflows": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_gorax_gorax_internal_analytics.TopWorkflows": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "workflows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_gorax_gorax_internal_analytics.TopWorkflow"
                    }
                }
            }
        },
        "github_com_gorax_gorax_internal_analytics.WorkflowStats": {
            "type": "object",
            "properties": {
                "avgDurationMs": {
                    "type": "integer"
                },
                "cancelledCount": {
                    "type": "integer"
                },
                "executionCount": {
                    "type": "integer"
                },
                "failureCount": {
                    "type": "integer"
                },
                "lastExecutedAt": {
                    "type": "string"
                },
                "maxDurationMs": {
                    "type": "integer"
                },
                "minDurationMs": {
                    "type": "integer"
                },
                "pendingCount": {
                    "type": "integer"
                },
                "runningCount": {
                    "type": "integer"
                },
                "successCount": {
                    "type": "integer"
                },
                "successRate": {
                    "type": "number"
                },
                "totalDurationMs": {
                    "type": "integer"
                },
                "workflowId": {
                    "type": "string"
                },
                "workflowName": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionCookie": {
            "description": "Ory Kratos session cookie (production mode)",
            "type": "apiKey",
            "name": "ory_kratos_session",
            "in": "cookie"
        },
        "TenantID": {
            "description": "Tenant identifier for multi-tenant isolation",
            "type": "apiKey",
            "name": "X-Tenant-ID",
            "in": "header"
        },
        "UserID": {
            "description": "User identifier (development mode only)",
            "type": "apiKey",
            "name": "X-User-ID",
            "in": "header"
        }
    },
...
}